separate component for model choosing

+++ Lib/Entities/Prediction.cs
@@ -3,6 +3,6 @@
     public class Prediction
     {
         public string Word = "";
-        public float Score = 0.0f;
+        public double Score = 0.0f;
     }
 }
